Bug fixed (player never dies) - Changed Player.prototype.kill and the way of calling "game over" in Controller.js

+++ Game/scripts/Controller.js
@@ -69,7 +69,11 @@ define(['game'], function (game) {
         }
 
         if (this.player.graphics.y === 464) {
-            this.player.kill();
+            if (this.player.lives) {
+                this.player.kill();
+            } else {
+                game.state.start('game-over', true, false, this.player);
+            }
         }
     };
